requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==13.0
twilio==8.10.0
google-genai==0.3.0
numpy==1.24.3
python-dotenv==1.0.0
